requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-database-url==3.1.0
Django==5.2.11
djangorestframework==3.16.1
idna==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
